# Improve performance of CompositePropertySource#getPropertyNames