--- a/spring-core/spring-core.gradle
+++ b/spring-core/spring-core.gradle
@@ -66,6 +66,7 @@ dependencies {
 	testFixturesImplementation("org.junit.jupiter:junit-jupiter-params")
 	testFixturesImplementation("org.assertj:assertj-core")
 	testFixturesImplementation("org.xmlunit:xmlunit-assertj")
+    jmh("org.apache.commons:commons-lang3:3.12.0")
 }
 
 jar {