@pivotal-cla 2021-08-02
@Shawyeok Thank you for signing the [Contributor License Agreement](https://cla.pivotal.io/sign/spring?repositoryId=spring-projects/spring-framework&pullRequestId=27236)!