@Shawyeok 2021-08-02
Related https://github.com/ctripcorp/apollo/pull/3816#discussion_r667444749