@Shawyeok 2023-08-27
> @Shawyeok thank you for making your first contribution to Spring Framework.

It's great to see it get merged even two years later.